Moved SearchModel.js, updated imports.
